ai-custom-voice-assistant/requirements.txt
# Requirements for AI Assistant PoC
openai>=1.0.0
python-dotenv
SpeechRecognition
pyttsx3
PyAudio
